Crate couchdb

The couchdb library provides types for working with CouchDB.

The couchdb library is not a CouchDB client. Rather, it makes it easier for applications to communicate with a CouchDB server using existing HTTP client libraries (such as hyper and reqwest).

The couchdb library is a toolkit, not a framework. Applications may opt in to using as much or as little of the library as makes the most sense.

Prerequisites

The application programmer is familiar with CouchDB and its API.

Though the couchdb library aims to be easy to use, it does not aim to teach programmers about CouchDB or how to use the CouchDB API. For more information about CouchDB, consult its documentation.

Remarks

The CouchDB API, like most HTTP interfaces, uses a lot of stringly types and requires client applications to do a lot of text-parsing and text-formatting. The couchdb library makes working with these stringly types easier.

In earlier versions, the couchdb library provided a fledgling CouchDB client for communicating with a CouchDB server, but now the library is purely a passive collection of types, as well as testing tools, that's intended to be used in conjunction with other HTTP libraries, such as hyper or reqwest.

Reexports

pub use attachment::Attachment;

pub use path::*;

Modules

attachment
The attachment module provides types for working with CouchDB document attachments.

path
The path module provides types for identifying databases, documents, etc.

testing
The testing module provides tools for applications to test their use of CouchDB.

Structs

Database
Database contains the content of a database resource.

Nok
Nok contains the content of an error response from the CouchDB server.

Revision
Revision contains a document revision.

Root
Root contains the content of a CouchDB server's root resource.

Vendor
Vendor contains information about a CouchDB server vendor.

Version
Version is a string specifying a version.

Enums

Error
Error is the principal type of the couchdb crate.