Module couchdb::action

The action module defines action types.

An action is an HTTP method paired with a CouchDB server resource—e.g., GET /db and PUT /db/doc.

Each action has a corresponding type—e.g., GetDatabase and PutDocument—used for performing that action.

Applications construct actions by calling the appropriate Client method—e.g., get_database and put_document.

CouchDB API coverage

This table shows in detail what parts of the CouchDB API this crate supports.

URI path or other feature 	Method 	Action type 	Header, query parameter, or other sub-feature 	Supported?

Basic Authentication 	No

/ 	GET 	GetRoot 	 	Yes

/_active_tasks 	GET 	 	 	No

/_all_dbs 	GET 	GetAllDatabases 	 	Yes

/_db_updates 	GET 	 	 	No

/_log 	GET 	 	 	No

/_replicate 	POST 	 	 	No

/_restart 	POST 	 	 	No

/_stats 	GET 	 	 	No

/_utils/ 	GET 	 	 	No

/_uuids/ 	GET 	 	 	No

/favicon.ico/ 	GET 	 	 	No

/_session 	POST 	 	 	No

GET 	 	 	No

DELETE 	 	 	No

/_config 	GET 	 	 	No

/_config/section 	GET 	 	 	No

/_config/section/key 	GET 	 	 	No

PUT 	 	 	No

DELETE 	 	 	No

/db 	HEAD 	HeadDatabase 	 	Yes

GET 	GetDatabase 	 	Yes

PUT 	PutDatabase 	 	Yes

DELETE 	DeleteDatabase 	 	Yes

POST 	PostDatabase 	 	Yes

X-Couch-Full-Commit 	No

?batch 	No

/db/_all_docs 	GET 	 	 	No

POST 	 	 	No

/db/_bulk_docs 	POST 	 	 	No

/db/_changes 	GET 	GetChanges 	 	Yes

Last-Event-Id 	No

?att_encoding_info 	No

?attachments 	No

?conflicts 	No

?descending 	No

?doc_ids 	No

?feed 	Yes

?filter 	No

?heartbeat 	Yes

?include_docs 	No

?last-event-id 	No

?limit 	No

?since 	Yes

?style 	No

?timeout 	Yes

?view 	No

/db/_compact 	POST 	 	 	No

/db/_compact/design-doc 	POST 	 	 	No

/db/_ensure_full_commit 	POST 	 	 	No

/db/_view_cleanup 	POST 	 	 	No

/db/_security 	GET 	 	 	No

PUT 	 	 	No

/db/_temp_view 	POST 	 	 	No

/db/_purge 	POST 	 	 	No

/db/_missing_revs 	POST 	 	 	No

/db/_revs_diff 	POST 	 	 	No

/db/_revs_limit 	GET 	 	 	No

PUT 	 	 	No

/db/doc

/db/_design/design-doc

/db/_local/id

HEAD 	HeadDocument 	 	Yes

If-None-Match 	Yes

GET 	GetDocument 	 	Yes

Multi-part attachments 	No

If-None-Match 	Yes

?att_encoding_info 	No

?attachments 	Yes

?atts_since 	No

?conflicts 	No

?deleted_conflicts 	No

?latest 	No

?local_seq 	No

?meta 	No

?open_revs 	No

?rev 	Yes

?revs 	No

?revs_info 	No

PUT 	PutDocument 	 	Yes

Multi-part attachments 	No

If-Match 	Yes

X-Couch-Full-Commit 	No

?batch 	No

DELETE 	DeleteDocument 	 	Yes

If-Match 	Yes

X-Couch-Full-Commit 	No

?rev 	Yes

?batch 	No

COPY 	 	 	No

/db/doc/attachment

/db/doc/_design/design-doc/attachment

HEAD 	 	 	No

GET 	 	 	No

PUT 	 	 	No

DELETE 	 	 	No

/db/_design/design-doc/_info 	GET 	 	 	No

/db/_design/design-doc/_view/view-name 	GET 	GetView 	 	Yes

?att_encoding_info 	No

?attachments 	No

?conflicts 	No

?descending 	No

?endkey 	Yes

?endkey_doc 	No

?group 	No

?group_level 	No

?include_docs 	No

?inclusive_end 	No

?key 	No

?limit 	No

?reduce 	Yes

?skip 	No

?stale 	No

?startkey 	Yes

?startkey_docid 	No

?update_seq 	No

PUT 	 	 	No

/db/_design/design-doc/_show/show-name 	GET 	 	 	No

PUT 	 	 	No

/db/_design/design-doc/_show/show-name/doc-id 	GET 	 	 	No

PUT 	 	 	No

/db/_design/design-doc/_list/list-name/view-name 	GET 	 	 	No

PUT 	 	 	No

/db/_design/design-doc/_list/list-name/other-ddoc/view-name 	GET 	 	 	No

PUT 	 	 	No

/db/_design/design-doc/_update/update-name 	POST 	 	 	No

/db/_design/design-doc/_update/update-name/doc-id 	PUT 	 	 	No

/db/_design/design-doc/_rewrite/path 	any 	 	 	No

Structs

DeleteDatabase
Action to delete a database.

DeleteDocument
Action to delete a document.

GetAllDatabases
Action to get all database names.

GetChanges
Action to get changes made to documents in a database.

GetDatabase
Action to get database meta-information.

GetDocument
Action to get document meta-information and application-defined content.

GetRoot
Action to get meta-information about the CouchDB server.

GetView
Action to execute a view.

HeadDatabase
Action to check whether a database exists.

HeadDocument
Action to check whether a document exists.

PostDatabase
Action to create a document.

PutDatabase
Action to create a database.

PutDocument
Action to create or update a document.

Traits

ChangeHandler
A change handler receives a single change result.