Struct couchdb::action::GetRoot

pub struct GetRoot<'a> { /* fields omitted */ }

Action to get meta-information about the CouchDB server.

Errors

All errors that occur as a result of executing this action are private.

Examples

The following example shows how to get the version of the CouchDB server.

extern crate couchdb;

let client = couchdb::Client::new("http://couchdb-server:5984").unwrap();
let root = client.get_root().run().unwrap();
assert_eq!("1.6.1", root.version.to_string());

Methods

impl<'a> GetRoot<'a>

fn run(self) -> Result<Root, Error>

Sends the action request and waits for the response.