Struct couchdb::action::PostDatabase

pub struct PostDatabase<'a, P, T> where P: IntoDatabasePath, T: 'a + Serialize {
    // some fields omitted
}

Action to create a document.

Return

This action returns the new document's revision and id.

Errors

The following are some of the errors that may occur as a result of executing this action:

Error::NotFound: The database does not exist.

Error::Unauthorized: The client is unauthorized.

Methods

impl<'a, P: IntoDatabasePath, T: 'a + Serialize> PostDatabase<'a, P, T>

fn run(self) -> Result<(DocumentId, Revision), Error>

Sends the action request and waits for the response.