Struct couchdb::command::GetAllDatabases

pub struct GetAllDatabases<'a> {
    // some fields omitted
}

Command to get a vector of all database names.

Methods

impl<'a> GetAllDatabases<'a>

fn new(client_state: &ClientState) -> GetAllDatabases

fn run(self) -> Result<Vec<String>, Error>

Send the command request and wait for the response.