Struct couchdb::command::GetDocument

pub struct GetDocument<'a, P, T> where P: IntoDocumentPath, T: Deserialize {
    // some fields omitted
}

Command to get document meta-information and application-defined content.

Return

This command returns an Option type. The return value is None if the command specifies a revision and the document hasn't been modified since that revision. Otherwise, the return value is Some and contains the document meta-information and application-defined content.

Errors

The following are some of the errors that may occur as a result of executing this command:

Error::NotFound: The document does not exist.

Error::Unauthorized: The client is unauthorized.

Methods

impl<'a, P: IntoDocumentPath, T: Deserialize> GetDocument<'a, P, T>

fn if_none_match(self, rev: &'a Revision) -> Self

Sets the If-None-Match header.

fn run(self) -> Result<Option<Document<T>>, Error>

Sends the command request and waits for the response.