Struct couchdb::command::GetView

pub struct GetView<'a, K, V> where K: Deserialize, V: Deserialize {
    // some fields omitted
}

Command to run a view.

Methods

impl<'a, K, V> GetView<'a, K, V> where K: Deserialize + Serialize, V: Deserialize

fn reduce(self, v: bool) -> Self

Set whether to run the reduce view function.

fn endkey(self, key: K) -> Self

Set the minimum key for rows contained within the result.

fn startkey(self, key: K) -> Self

Set the maximum key for rows contained within the result.

fn run(self) -> Result<ViewResult<K, V>, Error>

Send the command request and wait for the response.

Errors

Note: Other errors may occur.

Error::InternalServerError: An error occurred when executing the view.

Error::NotFound: The view does not exist.

Error::Unauthorized: The client is unauthorized.