Enum couchdb::Error

pub enum Error {
    DatabaseExists {
        response: ServerErrorResponse,
    },
    Decode {
        // some fields omitted
    },
    DocumentConflict {
        response: ServerErrorResponse,
    },
    Encode {
        // some fields omitted
    },
    InternalServerError {
        response: ServerErrorResponse,
    },
    InvalidDatabaseName {
        response: ServerErrorResponse,
    },
    InvalidRequest {
        response: ServerErrorResponse,
    },
    NotFound {
        response: Option<ServerErrorResponse>,
    },
    Transport {
        // some fields omitted
    },
    Unauthorized {
        error: String,
        reason: String,
    },
    UriParse {
        // some fields omitted
    },
    // some variants omitted
}

Principal error type.

The public API guarantees only the non-hidden variants and non-hidden variant fields. Do not match against hidden variants or hidden variant fields.

Variants

DatabaseExists

The database already exists.

Fields

response

Decode

JSON-decoding error.

Fields

DocumentConflict

The client request conflicts with an existing document.

Fields

response

Encode

JSON-encoding error.

Fields

InternalServerError

An internal server error occurred.

Fields

response

InvalidDatabaseName

The database name is invalid.

Fields

response

InvalidRequest

The client request is invalid.

Fields

response

NotFound

The resource does not exist.

Fields

response

CouchDB server response.

In case of a HEAD request, the response value is None (because the server doesn't send response content for HEAD requests). Otherwise, the response value is Some.

Transport

HTTP-transport error.

Fields

Unauthorized

The client is unauthorized to carry out the operation.

Fields

error

Error string returned by CouchDB Server.

reason

Reason string returned by CouchDB Server.

UriParse

URI-parse error.

Fields