Struct couchdb::path::DatabaseName

pub struct DatabaseName(_);

DatabaseName is a single URL path segment that specifies the name of a database.

For example, given the document path /db/_design/doc, the database name is db.

For more information about path-related types, see the module-level documentation.

Methods

impl DatabaseName

fn new<T: Into<String>>(s: T) -> Self

Constructs a new database name.

fn into_string(self) -> String

Converts the database name into a string.

impl DatabaseName

fn with_document_id<T: Into<DocumentId>>(self, doc_id: T) -> DocumentPath

Joins the database name with a document id to construct a document path.

fn with_design_document_id<T: Into<DesignDocumentId>>(
self,
ddoc_id: T
) -> DesignDocumentPath

Joins the database name with a design document id to construct a design document path.

fn into_database_path(self) -> DatabasePath

Converts the database name into a database path.