Struct couchdb::path::DesignDocumentId

pub struct DesignDocumentId(_);

DesignDocumentId comprises URL path segments that, together, identify a design document.

For example, given the document path /db/_design/doc, the design document id is _design/doc.

DesignDocumentId is a special form of DocumentId. All design document ids are document ids, but not all document ids are design document ids.

For more information about path-related types, see the module-level documentation.

Methods

impl DesignDocumentId

fn parse(s: &str) -> Result<Self, Error>

Tries to construct a design document id from a string.

fn into_string(self) -> String

Converts the design document id into a String.

fn into_document_id(self) -> DocumentId

Converts the design document id into a general document id.

fn into_design_document_name(self) -> DesignDocumentName

Converts the design document id into a design document name.