Struct couchdb::ChangeItemBuilder

pub struct ChangeItemBuilder {
    // some fields omitted
}

Builder for constructing a change item.

Methods

impl ChangeItemBuilder

fn new(rev: Revision) -> Self

Constructs a new change item builder.

The change item contained within the builder has the given revision.

fn new_from_rev_str<R: AsRef<str>>(rev: R) -> Self

Constructs a new change item builder.

The change item contained within the builder is assigned a revision equivalent to the given revision string.

Panics

Panics if the revision string is invalid.

fn unwrap(self) -> ChangeItem

Returns the change item contained within the builder.