Struct couchdb::ChangesBuilder

pub struct ChangesBuilder { /* fields omitted */ }

Builder for constructing a change list.

Methods

impl ChangesBuilder

fn new(last_seq: u64) -> Self

Constructs a new change list builder.

The change list contained within the builder has the given sequence number and no change results.

fn unwrap(self) -> Changes

Returns the change list contained within the builder.

fn push_result(self, result: ChangeResult) -> Self

Adds a change result to the change list contained within the builder.

fn build_result<I, F>(self, seq: u64, doc_id: I, f: F) -> Self where I: Into<DocumentId>,
F: FnOnce(ChangeResultBuilder) -> ChangeResultBuilder

Builds and adds a change result to the change list contained within the builder.