Struct couchdb::Database

pub struct Database {
    pub committed_update_seq: u64,
    pub compact_running: bool,
    pub db_name: DatabaseName,
    pub disk_format_version: i32,
    pub data_size: u64,
    pub disk_size: u64,
    pub doc_count: u64,
    pub doc_del_count: u64,
    pub instance_start_time: u64,
    pub purge_seq: u64,
    pub update_seq: u64,
    // some fields omitted
}

Database contains the content of a database resource.

Database has public members instead of accessor methods because there are no invariants restricting the data.

Database implements Deserialize.

Remarks

An application may obtain a database resource by sending an HTTP request to GET /{db}.

Compatibility

Database contains a dummy private member in order to prevent applications from directly constructing a Database instance. This allows new fields to be added to Database in future releases without it being a breaking change.

Fields
committed_update_seq: u64 compact_running: bool db_name: DatabaseName disk_format_version: i32 data_size: u64 disk_size: u64 doc_count: u64 doc_del_count: u64 instance_start_time: u64 purge_seq: u64 update_seq: u64