Struct couchdb::DesignBuilder

pub struct DesignBuilder {
    // some fields omitted
}

Builder for constructing a design document.

Examples

use couchdb::{DesignBuilder, ViewFunction};
let foo = ViewFunction {
              map: "function(doc) { emit(doc.name, doc.foo); }".to_string(),
              reduce: Some("_sum".to_string()),
          };
let bar = ViewFunction {
              map: "function(doc) { emit(doc.name, doc.bar); }".to_string(),
              reduce: None,
          };
let design = DesignBuilder::new()
                 .insert_view("foo", foo.clone())
                 .insert_view("bar", bar.clone())
                 .unwrap();
assert_eq!(foo, *design.views.get("foo").unwrap());
assert_eq!(bar, *design.views.get("bar").unwrap());

Methods

impl DesignBuilder

fn new() -> Self

Construct a builder containing an empty design document.

fn unwrap(self) -> Design

Return the design document contained within the builder.

fn insert_view<T, U>(self, view_name: T, view_function: U) -> Self where T: Into<String>, U: Into<ViewFunction>

Add a view function to the design document contained within the builder.

If the design document already contains a view with the same name then the new view function will replace the existing view function.