Crate couchdb

The couchdb crate is a thin wrapper around the CouchDB API, providing low-level access to individual CouchDB commands—e.g., PUT database, GET document, etc. The goal is for the crate to deal with the menial task of sending HTTP requests and receiving HTTP responses and allow application writers to focus on their business logic.

This documentation has been written with the assumption that the reader is familiar with the CouchDB API. Descriptions of types, methods, etc. in the couchdb crate should provide just enough information for the reader to map the crate's concepts onto the CouchDB API and then use the CouchDB documentation to fill in the remaining details. Most names in the crate are identical to the names used in the CouchDB API so as to make this mapping straightforward.

The couchdb crate provides applications with type-safety beyond working with raw strings. Applications get and put documents as structured types. Other types, such as revisions and views, are strongly typed as well.

To deal with JSON, the couchdb crate relies on Serde and its traits serde::Serialize and serde::Deserialize. These traits are fundamental to the crate's API—they are not mere implementation details. As such, applications must use these traits when working with documents, views, or any type that is JSON-encoded in the CouchDB API.

Examples

Here are some examples of how to use the couchdb crate.

Example #1: Put a document

This example creates a database named “baseball” and puts a single document into it.

extern crate couchdb;

#[derive(Serialize)]
struct Player {
    name: String,
    hits: u32,
    home_runs: u32,
}

let client = couchdb::Client::new("http://your_database/").unwrap();

client.put_database("baseball").run().unwrap();

let doc = Player {
    name: "Babe Ruth".to_string(),
    hits: 2873,
    home_runs: 714,
};

client.put_document("baseball/babe_ruth", &doc).run()
                                               .unwrap();

Example #2: Get a document

This example gets the document created in the previous example.

extern crate couchdb;

#[derive(Deserialize)]
struct Player {
  name: String,
  hits: u32,
  home_runs: u32,
}

let client = couchdb::Client::new("http://your_database/").unwrap();

let doc = client.get_document::<_, Player>("baseball/babe_ruth").run()
                                                                .unwrap()
                                                                .unwrap();

assert_eq!(doc.path, "baseball/babe_ruth".into());
assert_eq!(doc.content.name, "Babe Ruth".to_string());
assert_eq!(doc.hits, 2873);
assert_eq!(doc.home_runs, 714);

CouchDB API coverage

In the couchdb crate, the Client type is the principal type for communicating with a CouchDB server. All HTTP requests to the CouchDB server go through a Client instance.

This table maps each CouchDB API resource to the Client method that accesses that resource.

URI 	HTTP method 	Client method 	Description

/_all_dbs

GET 	get_all_databases 	Get list of all databases

/db

HEAD 	head_database 	Test whether a database exists

GET 	get_database 	Get meta-information about a database

PUT 	put_database 	Create a database

DELETE 	delete_database 	Delete a database

POST 	post_to_database 	Create a document

/db/doc

/db/_design/design-doc

/db/_local/id

HEAD 	head_document 	Test whether a document exists

GET 	get_document 	Get meta-information and application-defined content for a document

PUT 	put_document 	Create or update a document

DELETE 	delete_document 	Delete a document

/db/_design/design-doc/_view/view

GET 	get_view 	Execute a view

Modules

command
Sub-module for individual command types.

Structs

Client
CouchDB client.

Database
Database resource, as returned from a command to GET a database.

DatabasePath
Database path—i.e., database name.

Design
Content of a design document.

DesignBuilder
Builder for constructing a design document.

Document
Document, including both meta-information and application-defined content.

DocumentPath
Document path—i.e., a database path paired with a document id.

ErrorResponse
Response content from the CouchDB server in case of error.

Revision
Document revision.

Server
RAII wrapper for running a CouchDB server process.

ViewFunction
JavaScript map and reduce functions for a CouchDB view.

ViewPath
View path—i.e., a document path paired with a view name.

ViewResult
Response resulting from executing a view.

ViewRow
Single row contained within the response resulting from executing a view.

Enums

DocumentId
Document identifier.

Error
Principal error type.

Traits

IntoUrl
Trait for converting some types into a URI.

Type Definitions

ViewFunctionMap
Associative collection for view functions.