Crate couchdb

The couchdb crate is a thin wrapper around the CouchDB API, providing low-level access to individual CouchDB commands—e.g., PUT database, GET document, etc. The goal is for this crate to deal with the menial task of sending HTTP requests and receiving HTTP responses and allow application writers to focus on their business logic.

The couchdb crate provides applications with type-safety beyond working with raw strings. Applications get and put documents as structured types. Other types, such as revisions and views, are strongly typed as well.

To deal with JSON, the couchdb crate relies on Serde and its traits serde::Serialize and serde::Deserialize. These traits are fundamental to the couchdb crate's API—they are not mere implementation details. As such, applications must use these traits when working with documents, views, or any type that is JSON-encoded in the CouchDB API.

Examples

Here are some examples of how to use the couchdb crate.

Example #1: Put a document

This example creates a database named “baseball” and puts a single document into it.

extern crate couchdb;

#[derive(Serialize)]
struct Player {
    name: String,
    hits: u32,
    home_runs: u32,
}

let client = couchdb::Client::new("http://your_database/").unwrap();

client.put_database("baseball").run().unwrap();

let doc = Player {
    name: "Babe Ruth".to_string(),
    hits: 2873,
    home_runs: 714,
};

client.put_document("cats/babe_ruth", &doc).run()
                                           .unwrap();

Example #2: Get a document

This example gets the document created in the previous example.

extern crate couchdb;

#[derive(Deserialize)]
struct Player {
  name: String,
  hits: u32,
  home_runs: u32,
}

let client = couchdb::Client::new("http://your_database/").unwrap();

let doc = client.get_document::<_, Player>("baseball/babe_ruth").run()
                                                                .unwrap()
                                                                .unwrap();

assert_eq!(doc.path, "baseball/babe_ruth".into());
assert_eq!(doc.content.name, "Babe Ruth".to_string());
assert_eq!(doc.hits, 2873);
assert_eq!(doc.home_runs, 714);

Modules

command
Sub-module for individual command types.

Structs

Client
CouchDB client.

Database
Database resource, as returned from a command to GET a database.

DatabasePath
Database path—i.e., database name.

Design
Content of a design document.

DesignBuilder
Builder for constructing a design document.

Document
Document, including both meta-information and application-defined content.

DocumentPath
Document path—i.e., a database path paired with a document id.

ErrorResponse
Response content from the CouchDB server in case of error.

Revision
Document revision.

Server
RAII wrapper for running a CouchDB server process.

ViewFunction
JavaScript map and reduce functions for a CouchDB view.

ViewPath
View path—i.e., a document path paired with a view name.

ViewResult
Response resulting from executing a view.

ViewRow
Single row contained within the response resulting from executing a view.

Enums

DocumentId
Document identifier.

Error
Principal error type.

Traits

IntoUrl
Trait for converting some types into a URI.

Type Definitions

ViewFunctionMap
Associative collection for view functions.