Struct couchdb::DocumentPath

pub struct DocumentPath(_, _);

Document path—i.e., a database path paired with a document id.

A document path pairs a DatabasePath with a DocumentId—e.g., the db and docid parts in the HTTP request GET http://example.com:5984/db/docid.

DocumentPath provides additional type-safety over working with raw strings.

Methods

impl DocumentPath

fn new<T, U>(db_path: T, doc_id: U) -> Self where T: Into<DatabasePath>, U: Into<DocumentId>

fn into_uri(self, base_uri: Url) -> Url

Convert the DocumentPath into a URI.

fn database_path(&self) -> &DatabasePath

Return the database path part of the document path.

fn document_id(&self) -> &DocumentId

Return the document id part of the document path.