Struct couchdb::Revision

pub struct Revision { /* fields omitted */ }

Revision contains a document revision.

Revision stores a revision, which is a string that looks like 1-9c65296036141e575d32ba9c034dd3ee.

Revision can be parsed from a string via FromStr or the Revision::parse method.

Revision implements Deserialize and Serialize.

Remarks

A CouchDB document revision comprises a sequence number and an MD5 digest. The sequence number (usually) starts at 1 when the document is created and increments by one each time the document is updated. The digest is a hash of the document content, which the CouchDB server uses to detect conflicts.

Example

extern crate couchdb;

let rev = couchdb::Revision::parse("42-1234567890abcdef1234567890abcdef")
    .unwrap();

assert_eq!(rev.to_string(), "42-1234567890abcdef1234567890abcdef");

Methods

impl Revision

fn parse(s: &str) -> Result<Self, Error>

Constructs a new Revision from the given string.

The string must be of the form 42-1234567890abcdef1234567890abcdef.

fn sequence_number(&self) -> u64

Returns the sequence number part of the revision.

The sequence number is the 42 part of the revision 42-1234567890abcdef1234567890abcdef.