Struct couchdb::ViewResult

pub struct ViewResult<K, V> where K: Deserialize, V: Deserialize {
    pub total_rows: Option<u64>,
    pub offset: Option<u64>,
    pub rows: Vec<ViewRow<K, V>>,
}

Response resulting from executing a view.

Fields

total_rows

Total number of rows in the view, including those not contained within this result.

offset

Zero-based offset of the first row contained within this result.

The offset field specifies the number of rows in the view whose key is less than the key of the first row in the rows field.

rows

Rows contained within this result.

A view may have rows not contained within this result. For example, this may happen when using the startkey or endkey parameters when executing the view.